Move library code to sensible folder path.
