+++ src/main/scala/nl/tudelft/tydi_chisel/package.scala
@@ -1,4 +1,6 @@
-package object tydi_lib {
+package nl.tudelft
+
+package object tydi_chisel {
   val firNoOptimizationOpts: Array[String] = Array("-disable-opt", "-disable-all-randomization", "-strip-debug-info")
   val firNormalOpts: Array[String] = Array("-O=debug", "-disable-all-randomization", "-strip-debug-info")
   val firReleaseOpts: Array[String] = Array("-O=release", "-disable-all-randomization", "-strip-debug-info")
