+++ src/main/scala/nl/tudelft/tydi_chisel/utils/LastSeqProcessor.scala
@@ -1,8 +1,8 @@
-package tydi_lib.utils
+package nl.tudelft.tydi_chisel.utils
 
 import chisel3._
 import chisel3.util.{PriorityEncoder, Reverse, log2Ceil}
-import tydi_lib.TydiModule
+import nl.tudelft.tydi_chisel.TydiModule
 
 class LastSeqProcessor(val n: Int, val d: Int) extends TydiModule {
   val prevReduced: UInt = IO(Input(UInt(d.W)))
